ARROW-15599: [R] Convert a column as a sub-second timestamp from CSV file with the `T` col type option
The "T" option that can be set for `col_types` is mapped to `timestamp[s]` now, so it does not support sub-second timestamp.
Change the mapping to `timestamp[ns]`, which is the smallest unit, so that sub-second time can be read.
And add examples of specifying the column types to the reference.

Closes #12474 from eitsupi/fix-read-csv-timestamp

Lead-authored-by: SHIMA Tatsuya <ts1s1andn@gmail.com>
Co-authored-by: Jonathan Keane <jkeane@gmail.com>
Co-authored-by: Nicola Crane <thisisnic@gmail.com>
Signed-off-by: Jonathan Keane <jkeane@gmail.com>