ARROW-16305: [C++] Missed reference to ARROW_ENGINE during the rename
I had not thought to check .travis.yml.  The errors appeared to be completely unrelated and I'm still not entirely sure why we receive them.  However, in the past, we were not building parquet on the s390x build.  With ARROW-16158 we accidentally enabled parquet on this build.  This change should disable `ARROW_SUBSTRAIT` (and transitively, `ARROW_PARQUET`) so that the behavior is the same as it was before.

Closes #12979 from westonpace/bugfix/MINOR--missed-ref-to-arrow-engine

Authored-by: Weston Pace <weston.pace@gmail.com>
Signed-off-by: Weston Pace <weston.pace@gmail.com>