ARROW-16504 [Go][CSV] Add arrow.TimestampType support to the reader
There is already a helper to convert strings to arrow.Timestamp so incorporate this into the CSV reader.

The CSV files I am currently working with have RFC3339 timestamps so I followed some of the code JSON and stuck with millisecond default.

Was really easy to add this using the existing functions and structure.

Closes #13098 from wolfeidau/ARROW-16504-add-timestamp-support-to-reader

Authored-by: Mark Wolfe <mark@wolfe.id.au>
Signed-off-by: Matt Topol <zotthewizard@gmail.com>