MINOR: [C++] Move static declaration to non-static declaration to improve performance (#13222)
According to conbench there was a slight regression on #12957 .  Poking around a bit it seems that a static local variable is implemented using some kind of global lock (__cxa_guard_acquire / __cxa_guard_release).  On the other hand, constructing an empty shared_ptr is cheap (two pointers are set to 0).  So if we care about performance here we probably don't want `static`.  This may be what is causing the conbench issue.

Lead-authored-by: Weston Pace <weston.pace@gmail.com>
Co-authored-by: Antoine Pitrou <antoine@python.org>
Signed-off-by: Yibo Cai <yibo.cai@arm.com>