ARROW-16759: [Go] update testify to get security patch for gopkg.in/yaml.v3 (v7)
This PR updates the github.com/stretchr/testify dependency to get a security patch for gopkg.in/yaml.v3 which has a DoS exploit. See https://github.com/stretchr/testify/pull/1192 for more details.

I'm unsure how this project handles security patches for appears to be older versions. I'm here because I have dependencies that rely on v7, so that's what is bringing me here to make this very particular change. It looks like v6.0.0 and v6.0.1 tags exist, so I expect merging this here and tagging v7.0.1 would be the path forward. If not, let me know what would be preferred.

The linked Jira issue also calls out v8.0.0 as having the same vulnerability, but that would need to be addressed in it's own PR.

Closes #13322 from dominicbarnes/go-security-patch-testify

Authored-by: Dominic Barnes <dominic@dbarnes.info>
Signed-off-by: Matthew Topol <mtopol@factset.com>