ARROW-17087: [C++] Race condition in scanner test (#13651)
The ExecPlan::AddTask variant that takes a future was problematic because a future is not created until a task has already been started and, if the task group was already ended, it wasn't tracking the future.  So work would be started that would not be tracked anywhere and the plan was ending prematurely.  As a fix I changed that AddTask variant to BeginExternalTask.  This method will request permission to start the task before the future is created.  At some point the scanner should be moved to an exec plan instead of async generators and I think BeginExternalTask can go away.

Lead-authored-by: Weston Pace <weston.pace@gmail.com>
Co-authored-by: Antoine Pitrou <antoine@python.org>
Signed-off-by: Antoine Pitrou <antoine@python.org>